Make Block movable, align on size_t for size properties.
* Use size_t for dimensional properties to align with rest of code
  (and general container `.size()` types)
* Adopt MonoView for views
* Remove `Swap()` and `SetNumChannels()` methods to emphasize the
  implicit constness of the class (i.e. encourage setting dimensions
  only at construction time).

Bug: chromium:335805780

+++ modules/audio_processing/aec3/coarse_filter_update_gain_unittest.cc
@@ -78,8 +78,8 @@ void RunFilterUpdateTest(int num_blocks_to_process,
                   k) != blocks_with_saturation.end();
 
     // Create the render signal.
-    for (int band = 0; band < x.NumBands(); ++band) {
-      for (int channel = 0; channel < x.NumChannels(); ++channel) {
+    for (size_t band = 0; band < x.NumBands(); ++band) {
+      for (size_t channel = 0; channel < x.NumChannels(); ++channel) {
         RandomizeSampleVector(&random_generator, x.View(band, channel));
       }
     }
